Module hydro_test::cluster::paxos

Structs

Acceptor

Ballot

CorePaxos

LogValue

P2a

PaxosConfig

Proposer

Enums

CheckpointOrP2a

Traits

PaxosPayload

Functions

acceptor_p2

index_payloads

leader_election⚠

paxos_core⚠

Implements the core Paxos algorithm, which uses a cluster of propsers and acceptors to sequence payloads being sent to the proposers.

recommit_after_leader_election